fix #12 maybe
Several Chinese translation mods make extensive changes to the Text class. These seem to have bad interactions with my relatively minor change to record the font and size when they are set, so I will delay the modification of the Text class until the clock loads, and not have the constructor call SetFont or SetSize again, since they may be modified to have other behavior.